Add missed rcu_read_[un]lock() calls
The RCX code is written in assumption of our machine, which provides
total store order and running non-preemptible kernel.  Thus, we skipped
marking rcu read critical sections as rcu_read_[un]lock() are just no-op
and only make code long and complex.  This commit adds the rcu read
critical sections as the code is not for only our machine.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>